Remove semicola from SNOWHOUSE_ASSERT_THAT/AssertThat definitions
Note that this is actually a backwards incompatible change since we
now expect a semicolon after every use of these macros.

Since the old behavior (it works without semicolon) was not documented,
I refrain from documenting that change. However, according to
Semantic Versioning, this non-issue requires an increment of the
major version.