appveyor.yml: Run tests as test_script
Since the output of the tests changed, we have a line containing the word
"error" in the test output.  Since the tests are run within the build_script
of AppVeyor, this makes the build process think there was an error and
hence fails.  Extracting the tests to the test_script should fix the problem.