Make indenting widths within files more consistent
Note that this commit only makes the indenting consistent
to most of the other places in the same file. It is not a
consistent indenting over the full snowhouse codebase (which
is either 2 or 4 indenting spaces, and different behavior for
continuation).

We also stick to tab-less indenting here. Tabs must not replace
any number of spaces.