Change "type *foo" to "type* foo"
This style is used all over the snowhouse codebase except for
a small piece recently introduced by me.

The main advantage is that the pointer-* (and the reference-&)
belongs to the type (unless you use it to dereference a pointer
or want to get the address of a variable, of course).

The only disadvantage occurs when using commas in declaration,
for example,

	int* foo, bar;

makes one think that bar is also a pointer, but it is not.
However, commas in declaration for pointer types is bad style
either way, so we ignore that.