sequence_container_tests: Make insert_numbers() a template specialization
The specialized versions of insert_numbers() are overloaded functions
while the "general" version is a template. This works, but it is
weird and confusing (but I have a lack of experience if there are other
(dis)advantages with this... it probably does not matter because these
are static functions and not any library code).

This commit realizes the specialized versions as template specializations
to make this consistent.